Restore profiling macros
To allow disabling Tracy entirely, something need to replace its macros with empty ones.
It's possible to do so by keeping the same macros, but I'd also like to make it a little clearer what is and isn't profiling-related, and make it easier to replace if need be.
So I'm reviving the previously commented out in-between macros used for this.